Added ChatList.js. Code view updated
