+++ web-components/src/index.html
@@ -5,10 +5,10 @@
     <title>Create message</title>
 </head>
 <body>
-    <div class="chat-window">
-        <chat-header></chat-header>
-        <message-form action="/">
-        </message-form>
+    <div class="main-window">
+        <app-header></app-header>
+        <message-form action="/"></message-form>
+        <chats-list></chats-list>
     </div>
 </body>
 </html>
