+++ web-components/src/index.js
@@ -1,5 +1,6 @@
 import './index.css';
 
-import './components/ChatHeader';
+import './components/AppHeader';
 import './components/FormInput';
 import './components/MessageForm';
+import './components/ChatsList';
